Added deletion endpoints user, event, location and seat. Consolidated dtos.

+++ seat-reservation/src/main/java/com/fragment/seat_reservation/dto/SeatCreationDto.java
@@ -19,6 +19,6 @@ public class SeatCreationDto {
 
     private Integer seatCount;
 
-    private Integer location;
+    private Long location;
 
 }
